Re-do as mdbook
