+++ README.md
@@ -1,11 +1,7 @@
 # Lifetime variance in Rust
 
 This worked example covers the basics of *variance* in Rust, as it applies to
-lifetimes, written in a literate programming style.
-
-This is a working Rust crate. Clone this repo and run `cargo build` to play with it.
-
-**Head on over to [`src/lib.rs`](src/lib.rs) to begin!**
+lifetimes.
 
 ## License
