+++ additional/ferris.css
@@ -1,0 +1,23 @@
+body.light .does_not_compile,
+body.rust .does_not_compile {
+  background: #fff1f1;
+}
+
+body.coal .does_not_compile,
+body.navy .does_not_compile,
+body.ayu .does_not_compile {
+  background: #501f21;
+}
+
+.ferris {
+  position: absolute;
+  z-index: 99;
+  right: 5px;
+  top: 30px;
+	width: 10%;
+	height: auto;
+}
+
+.ferris-explain {
+  width: 100px;
+}
