+++ code/ch01-01-building-an-intuition/cell-counterexample/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "cell-counterexample"
+version = "0.1.0"
+authors = ["Rain <rain@sunshowers.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
