+++ code/ch01-01-building-an-intuition/cell-shortener/src/main.rs
@@ -1,0 +1,11 @@
+#![allow(dead_code)]
+fn main() {}
+
+use std::cell::Cell;
+
+// ANCHOR: all
+fn cell_shortener<'a, 'b>(s: &'a Cell<&'static str>) -> &'a Cell<&'b str> {
+    s
+}
+
+// ANCHOR_END: all
