+++ code/ch01-01-building-an-intuition/fn-ptr-lengthener/src/main.rs
@@ -1,0 +1,8 @@
+#![allow(dead_code)]
+fn main() {}
+
+// ANCHOR: all
+fn fn_ptr_lengthener<'a>(f: fn(&'a str) -> ()) -> fn(&'static str) -> () {
+    f
+}
+// ANCHOR_END: all
