+++ code/ch01-01-building-an-intuition/lifetime-lengthener/src/main.rs
@@ -1,0 +1,9 @@
+#![allow(dead_code)]
+fn main() {}
+
+// ANCHOR: all
+fn lifetime_lengthener<'a>(s: &'a str) -> &'static str {
+    s
+}
+
+// ANCHOR_END: all
