+++ code/ch01-01-building-an-intuition/lifetime-shortener/src/main.rs
@@ -1,0 +1,8 @@
+#![allow(dead_code)]
+fn main() {}
+
+// ANCHOR: all
+fn lifetime_shortener<'a>(s: &'static str) -> &'a str {
+    s
+}
+// ANCHOR_END: all
