+++ code/ch01-02-formalizing-variance/outlives-example/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "outlives-example"
+version = "0.1.0"
+authors = ["Rain <rain@sunshowers.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
