+++ code/ch01-02-formalizing-variance/outlives-example/src/main.rs
@@ -1,0 +1,9 @@
+#![allow(dead_code)]
+fn main() { }
+
+// ANCHOR: all
+struct OutlivesExample<'a, 'b: 'a> {
+    a_str: &'a str,
+    b_str: &'b str,
+}
+// ANCHOR_END: all
