+++ code/ch01-03-conflicts-and-type-parameters/two-spots/src/main.rs
@@ -1,0 +1,11 @@
+#![allow(dead_code)]
+fn main() {}
+
+use std::cell::Cell;
+
+// ANCHOR: all
+struct TwoSpots<'a> {
+    foo: &'a str,
+    bar: Cell<&'a str>,
+}
+// ANCHOR_END: all
