+++ code/ch01-04-variance-in-practice/overview/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "overview"
+version = "0.1.0"
+authors = ["Rain <rain@sunshowers.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
