+++ src/SUMMARY.md
@@ -1,0 +1,9 @@
+# Summary
+
+- [Lifetime Variance Example](./ch01-00-lifetime-variance-example.md)
+  - [Building an intuition](./ch01-01-building-an-intuition.md)
+  - [Formalizing variance](./ch01-02-formalizing-variance.md)
+  - [Conflicts and type parameters](./ch01-03-conflicts-and-type-parameters.md)
+  - [Variance in practice](./ch01-04-variance-in-practice.md)
+  - [Epilogue](./ch01-05-epilogue.md)
+- [Acknowledgements](./ch02-00-acknowledgements.md)
