+++ src/ch01-00-lifetime-variance-example.md
@@ -1,0 +1,3 @@
+# Lifetime Variance Example
+
+This worked example covers the basics of *variance* in Rust, as it applies to lifetimes.
