+++ src/ch01-05-epilogue.md
@@ -1,0 +1,6 @@
+# Epilogue
+
+Hope this made you feel more confident using lifetimes in your Rust code!
+They're a very powerful way to write safe, blazing fast code. But variance can
+often cause obscure issues in practice -- knowledge of how it works is key to
+using lifetimes effectively.
