+++ src/ch02-00-acknowledgements.md
@@ -1,0 +1,4 @@
+# Acknowledgements
+Thanks to the following people for their feedback:
+  * Nikolai Vazquez ([Twitter](https://twitter.com/NikolaiVazquez), [GitHub](https://github.com/nvzqz))
+  * Inanna Malick ([Twitter](https://twitter.com/inanna_malick), [GitHub](https://github.com/inanna-malick))
